perf: convert ApiDocsView to mostly-Astro to drop ~128 kB JS bundle (#208)
* perf: convert ApiDocsView to mostly-Astro to drop ~128 kB JS bundle (#186)

Replace the ApiDocsView React island with static Astro HTML:
- All prose, tables, and code blocks are now server-rendered
- src/components/code-block.astro: Shiki syntax highlighting via
  astro:components <Code>, copy button handled by vanilla JS script
- src/components/api-key-reveal.tsx: tiny React island (show/hide + copy)
  for the API key widget only
- Remove prism-react-renderer dependency

The api-docs-view.js chunk (126 kB raw / ~34 kB gzip) is eliminated.
baseUrl is now rendered server-side from Astro.url.origin.
