+++ package.json
@@ -46,7 +46,6 @@
     "lucide-react": "^0.548.0",
     "oxfmt": "^0.50.0",
     "oxlint": "^1.65.0",
-    "prism-react-renderer": "^2.4.1",
     "radix-ui": "^1.4.3",
     "react": "^19.2.0",
     "react-day-picker": "^9.13.0",
