+++ src/components/code-block.astro
@@ -1,0 +1,92 @@
+---
+import { Code } from "astro:components";
+interface Props {
+  code: string;
+  lang?: string;
+  title?: string;
+}
+const { code, lang = "bash", title } = Astro.props;
+---
+
+<div
+  class="relative group rounded-xl overflow-hidden border border-gray-800 shadow-lg"
+>
+  {
+    title && (
+      <div class="flex items-center justify-between px-4 py-2 bg-gray-800 border-b border-gray-700">
+        <div class="flex items-center gap-2">
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="14"
+            height="14"
+            viewBox="0 0 24 24"
+            fill="none"
+            stroke="currentColor"
+            stroke-width="2"
+            stroke-linecap="round"
+            stroke-linejoin="round"
+            class="text-gray-400"
+          >
+            <polyline points="4 17 10 11 4 5" />
+            <line x1="12" x2="20" y1="19" y2="19" />
+          </svg>
+          <span class="text-sm font-medium text-gray-300">{title}</span>
+        </div>
+        <span class="text-xs text-gray-500 uppercase tracking-wider">{lang}</span>
+      </div>
+    )
+  }
+  <div class="[&_pre]:p-4 [&_pre]:overflow-x-auto [&_pre]:text-sm [&_pre]:leading-relaxed [&_pre]:m-0">
+    <Code code={code.trim()} lang={lang} theme="night-owl" />
+  </div>
+  <button
+    class="copy-btn absolute top-2 right-2 p-2 rounded-lg bg-gray-700/50 opacity-0 group-hover:opacity-100 transition-all duration-200 hover:bg-gray-600 text-gray-300 hover:text-white"
+    data-code={code.trim()}
+    aria-label="Copy code"
+  >
+    <svg
+      class="icon-copy"
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      stroke-width="2"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+    >
+      <rect width="14" height="14" x="8" y="8" rx="2" ry="2" />
+      <path d="M4 16c-1.1 0-2-.9-2-2V4c0-1.1.9-2 2-2h10c1.1 0 2 .9 2 2" />
+    </svg>
+    <svg
+      class="icon-check hidden"
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      stroke-width="2"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+    >
+      <path d="M20 6 9 17l-5-5" />
+    </svg>
+  </button>
+</div>
+
+<script>
+  document.querySelectorAll<HTMLButtonElement>(".copy-btn").forEach((btn) => {
+    btn.addEventListener("click", async () => {
+      const code = btn.dataset.code ?? "";
+      await navigator.clipboard.writeText(code);
+      btn.querySelector(".icon-copy")?.classList.add("hidden");
+      btn.querySelector(".icon-check")?.classList.remove("hidden");
+      setTimeout(() => {
+        btn.querySelector(".icon-copy")?.classList.remove("hidden");
+        btn.querySelector(".icon-check")?.classList.add("hidden");
+      }, 2000);
+    });
+  });
+</script>
